requirements.txt
oracledb>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlparse>=0.4.4
ujson>=5.8.0
pyodbc>=4.0.0
typing_extensions>=4.5.0
